191022 | relationship

+++ articles/views.py
@@ -5,6 +5,10 @@ from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from .forms import ArticleForm, CommentForm
 from .models import Article, Comment
+# from accounts.models import User
+from django.contrib.auth import get_user_model
+from django.conf import settings
+from django.http import HttpResponseForbidden
 
 # Create your views here.
 @require_GET
@@ -34,6 +38,7 @@ def create(request):
     if request.method == 'POST':
     # POST 요청 -> 검증 및 저장
         article_form = ArticleForm(request.POST, request.FILES)
+        
         # embed()
         if article_form.is_valid():
         # 검증에 성공하면 저장하고,
@@ -41,7 +46,9 @@ def create(request):
             # new_content = article_form.cleaned_data.get('content')
             # article = Article(title=new_title, content=new_content)
             # article.save()
-            article = article_form.save()
+            article = article_form.save(commit=False)
+            article.user = request.user
+            article.save()
             # redirect
             return redirect('articles:detail', article.pk )
     else:
@@ -69,14 +76,16 @@ def detail(request, article_pk):
     return render(request, 'articles/detail.html', context)
     
 @require_POST
+@login_required
 def delete(request, article_pk):
     # article = Article.objects.get(pk=article_pk)
     article = get_object_or_404(Article, pk=article_pk)
+    if article.user == request.user:
     # if request.method == "POST":
-    article.delete()
-    return redirect('articles:index')
-    # else:
-    #     return redirect('articles:derail', article_pk)
+        article.delete()
+        return redirect('articles:index')
+    else:
+        return HttpResponseForbidden()
 
 # def edit(request, article_pk):
 #     if request.method == 'GET':
@@ -95,23 +104,27 @@ def delete(request, article_pk):
 def update(request, article_pk):
     # article = Article.objects.get(pk=article_pk)
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        article_form = ArticleForm(request.POST, instance=article)
-        # article = Article.objects.get(pk=article_pk)
-        # article.title = request.POST.get('title')
-        # article.content = request.POST.get('content')
-        # article.save()
-        if article_form.is_valid():
-            article = article_form.save()
-            return redirect('articles:detail', article.pk )
+    if request.user == article.user:
+        if request.method == 'POST':
+            article_form = ArticleForm(request.POST, instance=article)
+            # article = Article.objects.get(pk=article_pk)
+            # article.title = request.POST.get('title')
+            # article.content = request.POST.get('content')
+            # article.save()
+            if article_form.is_valid():
+                article = article_form.save()
+                return redirect('articles:detail', article.pk )
+        else:
+            article_form = ArticleForm(instance=article)
+        context = {
+            'article_form': article_form
+        }
+        return render(request, 'articles/form.html', context)
     else:
-        article_form = ArticleForm(instance=article)
-    context = {
-        'article_form': article_form
-    }
-    return render(request, 'articles/form.html', context)
+        return HttpResponseForbidden()
 
 @require_POST
+@login_required
 def comment_create(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     # 1. modelform에 사용자 입력값 넣고
@@ -123,6 +136,7 @@ def comment_create(request, article_pk):
         comment = comment_form.save(commit=False)
         # 3-2. FK 넣고 저장
         comment.article = article
+        comment.user = request.user
         comment.save()
     # 4. return redirect
     else:
@@ -139,8 +153,11 @@ def comment_create(request, article_pk):
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
+    if request.user == comment.user:
     # article_pk = comment.article_id
-    comment.delete()
+        comment.delete()
     # messages.error(request, '댓글이 삭제되었습니다.')
-    messages.add_message(request, messages.INFO, '댓글이 삭제되었습니다.' )
-    return redirect('articles:detail', article_pk)
+        messages.add_message(request, messages.INFO, '댓글이 삭제되었습니다.' )
+        return redirect('articles:detail', article_pk)
+    else:
+        return HttpResponseForbidden()
